+++ src/render/backends/mod.rs
@@ -6,6 +6,9 @@
 pub mod scene;
 pub mod sw;
 
+#[cfg(any(feature = "sdl-gpu", feature = "wgpu"))]
+mod lyon_path;
+
 #[cfg(feature = "sdl-gpu")]
 pub mod sdl_gpu;